+++ CHANGELOG.md
@@ -10,6 +10,8 @@
   to define an executable that must be in the `PATH` in order for it to work,
   and a command the user can use to install the executable if it doesn't exist
 * All built-in hooks will now fail if the required executable is not present
+* Fix bug where pre-commit hook would crash if user attempted to commit a
+  broken symlink
 
 ## 0.17.0